+++ admin/js/controller/keyManage.js
@@ -93,6 +93,8 @@ app.register.controller('Access', ['$scope', '$http', '$q', function($scope, $ht
             })(),
 
             delete_one: function (item) {
+                var r=confirm("确认删除access key "+ item.key_id +"吗？");
+                if (r===false) return;
                 $scope.access.data.delOneToken = Math.random();
                 $scope.aborter = $q.defer(),
                     $http.delete("http://121.41.72.231:5001/api/ivc/v1/access_keys/"+item.key_id, {
